add NVD3Chart.buildcontent()
NVD3Chart().buildcontent() constructs NVD3Chart().htmlcontent with no head,
body tags. To be useful probably need the attribute jquery_on_ready to be set,
so the chart is evaluated once the page is loaded.

This is useful if you are plugging NVD3Chart into another templating engine
(such as Tornado's) and simply require the chart and content.

Note: updated examples/demo_all.py to use the buildcontent() method.